Use database to hold pack store configuration
The pack stores now have a factory function that takes an enum and returns a
store implementation with a default configuration. The stores are configured
using JSON formatted properties. The store configuration is saved to the
database in JSON format as well, with the key being "store/<type>/<name>"
where the <name> is some name given to the store. Functions for finding
previously saved stores, and rebuilding them from the saved configuration,
are provided.

cargo test passes