Packs now can be stored in multiple locations
The saved pack record can have multiple "pack locations" which consist of
the store identifier, the remote bucket name, and the remote object name.
The stores are responsible for creating the location instance, since only
they know if it was necessary to use a different bucket or object name.

The store module now provides functions for storing a pack to a list of
stores, and returns the list of pack locations created by the stores.
Additionally, the module provides a function for taking a list of stores and
pack locations, from which it selects the best store for retrieving the
pack. First would be any local store, followed by "fast" remote stores (so
far they are all fast, since we do not yet have Glacier).

cargo test passes