Cache DB instances for each path
Once the last strong reference is dropped, the database will be closed. By
keeping a weak reference in a static cache, with the file path of the
database files as the key, we have a better chance of keeping a single
database instance opened when accessed from different threads. RocksDB
itself is thread-safe for get/put/write and the DB type implements Send and
Sync, should it should be safe to use in multiple threads.

cargo test passes