Replace gpgme with libsodium for encryption
The use of gpgme was not working properly when running in a container on the
Linux server (seemed like the passphrase prompt was not being invoked). In
the long run, this solution would have required including the gpg binary
along with the library.

The encryption is now handled using libsodium (via the Rust wrapper
sodiumoxide), which links with the application, and it works on Docker.

cargo test passes