Backup "starts" when the snapshot is started
Fix a mistake in the engine in which the "start backup" action was fired
after the snapshot was taken. The trouble is that a snapshot can take a long
time to build, and the supervisor starts another thread on top of the first
one, perhaps even repeatedly.

Add tests to ensure the supervisor properly determines whether a backup
should be running or not under various conditions.

cargo test passes