Use blob_uuid to shorten the computer ID
The UUID v5 is very useful for generating bucket names that conform to the
requirements of services such as Google Cloud, but they are a bit long. The
blob_uuid crate shortens them, and can convert back. So the computer ID is
the shortened form, both for storage in the database and display in the user
interface.

cargo test passes