Dataset form improvements
Use a text area for the stores value, and make it a little more clear that
the input can take a comma-separated list of values. Also fixed up the code
for splitting on commas by using Js functions.

Make the basepath and list of stores read-only when editing an existing
dataset. It would be rather horrible to allow the user to change these
values once there are snapshots.

cargo test passes