Ignore files changing during backup
There is no use in trying to detect a file changing between the time the
snapshot is taken and the time the file is packed. There will always be a
race condition, so just ignore that case altogether.

Turns out that removing a field from a struct and deserializing from the
database is not an issue, so no need to rewrite the database records.

cargo test passes