Handle file errors during backup
If a file goes missing (or has any other error) while attempting to add it
to a pack file (i.e. while finding chunks or simply copying the file into
the pack), record it as having a zero length.

Files with zero length are also handled correctly. They are stored in the
database and a chunk record is added that represents all empty files.

cargo test passes