Replace cron with an improved scheduling system
The cron crate was too limited for the purposes of scheduling not only a
time to start the backup, but a time to temporarily pause the process. That
is, running the backup only during a specific window of time. While that is
not entirely implemented yet, the code is in place. For now, only hourly,
daily, weekly, and monthly schedules are supported in the web interface. The
supervisor will not consider the end of the time range, if one were to be
provided somehow (e.g. via GraphQL).

cargo test passes