Rewrite most of the GraphQL resolvers
Employ the juniper object proc macro to define resolvers for some of the
core types, as well as the query and mutation resolvers. Reduced the code
considerably as a result. All of the extra GraphQL code is in the schema
module, with the exception of the easy derive macro and such for some of the
core types.

cargo test passes