Use chrono for any date/time stored in database
The use of SystemTime has rather limited usefulness since it is not related
to any particular time zone. It is the "system clock", whatever that means,
and thus is not reliable for persistent data storage.

Also, juniper supports the chrono values automatically, which saves the
trouble of serde on the values for GraphQL. Juniper formats the dates as a
string that JavaScript can parse.

cargo test passes