Change snapshot to have an unchanging "checksum"
Using the Snapshot.checksum() function was a mistake, as the snapshot
properties are changed over the course of the snapshot's lifetime. As a
result, the checksum changes, so the value stored in the dataset does not
match the one computed by the GraphQL resolver when retrieving the checksum
from the database.

cargo test passes