Use entities all the way down the stack
Use the entities when interacting with the use cases, repositories, and data
sources, pushing the "details" as far from the center (the "policy") as is
reasonable. In particular, the manner by which the pack stores were updated
via the data sources was leaking into the domain layer, which is quite
unnecessary. The detail of the pack store label having to be passed along
with the options should be confined to as small an area as possible.

flutter test passes