Start the Clean Architecture major rewrite
Not a change in functionality, and not really rewriting anything per se, but
rather organizing all of the code according to the Clean Architecture design
principles. A lot of code will move around and so nearly everything is
commented out until the move is complete.

With this first commit the Checksum and Chunk entities are defined, and the
repository interface and implementation for inserting and getting chunks has
been established. The data source is the RocksDB instance, as before.

cargo test passes