Split application into mulitple packages
The hope is that each package will build faster since they are separate
compilation units for the Rust compiler. Code changes are minimal, however
the pack store implementations have been split into a very thin adapter and
the actual pack store. The database code has moved to its own package, and
implements a trait defined in a new "database_core" package. For now, that
core package only defines a trait and nothing else, but someday it may be
useful to have it in place.

cargo test passes