fix: big grey rectangle issue in dataset form
The big grey rectangle in the dataset form when the start/stop times were
set was caused by integer math going wrong in the dart2js compiled code.
That is, with dartdevc there was no problem, but with the release build of
the application, the integer math was causing a mysterious error, which led
to a giant grey rectangle.

flutter test passes