fix: refine the scheduling controls
Only present the hourly and daily schedule frequencies, as the other options
do not have full configurability in the interface.

Selecting the daily frequency will make the start/stop time fields editable,
whereas for hourly they are made read-only.

Made some of the class methods to be static as they did not really have
anything to do with the widget instance.

flutter test passes