fix: track database "packs" for accurate pruning
The pack pruning was aggressively removing the database snapshots in the
pack store, since they were not being tracked in the database. Now they are
recorded using the key prefix "dbase/" and are formatted identically to
pack records. The use cases for finding and pruning now include these
additional records.

cargo test passes