Use same bucket name up to N times
Try to use the same bucket name for many pack files, rather than making a
new bucket name for each snapshot. The pack repository will produce the
bucket name on demand, generating a new name every 128 calls. The backup
process calls this function for each pack file that is about to be uploaded
to the pack store.

cargo test passes