feat: store very small files in tree record
Very small files, less than or equal to 80 bytes in length, will be stored
directly in the tree record as part of the tree reference. This is akin to
how symbolic links are handled. The size of 80 was chosen due to the fact
that the current file record serialized using CBOR is just over 80 bytes.
Hence, storing very small files directly in the tree will save space as we
avoid the database and tar file overhead.

Many test fixtures needed to be increased in size to avoid them being
treated as very small files and hence throwing off the test results.

cargo test passes