feat: support restoring symlinks, very small files
Changed the GraphQL API such that restore requests are now keyed by the tree
digest and the entry name within that tree to be restored. This allows for
restoring a symblink or very small file, neither of which have a file digest
value to use as the restore request parameter.

cargo/flutter test passes