fix: run pack store operations on separate thread
The pack stores that use asynchronous runtimes (google and minio) will have
their operations invoked on a separate thread. This allows the pack store
implementation to spawn a runtime for running futures, if necessary. Not all
pack stores do this as their underlying sources are synchronous.

Add a test that exercises the backup supervisor and the file restorer within
an async runtime (actix), with minio pack source, to ensure everything works
smoothly with regards to the async runtimes.

cargo test passes