fix: use firestore to rename database bucket
If the database bucket collides with another bucket that the given project
does not have permission to write to, then generate a new name instead. The
new name is mapped from the old name in Google Firestore, making it possible
for disaster recovery to find the correct bucket name when provided with the
original name (computer identifier).

cargo test passes