chore: arrange backup code for improved testing
Add a trait and implementation for what is essentially the function that
performs backups. This allows the backup supervisor code (now called the
"scheduler") to be tested more thoroughly via mock objects. A backup is
initiated by creating a request and sending it to the backup performer to
run the actual backup.

cargo test passes