feat: handle bucket collisions for pack stores
When storing pack files, as opposed to the database snapshots, detect and
work around bucket name collisions. That is, the cloud provider indicates
that the desired bucket name is already taken, and as such the pack repo
must generate a new name and try again. The pack store will return the
updated bucket name in the pack location struct.

cargo test passes