major: change from tar to 7z for pack files
Also the database snapshots are now compressed and archived using 7z. The
result seems to be good as the output from the unit tests is much smaller.

The pack file is written using 7z but still encrypted manually since we want
to control the salt that is used for improved security.

cargo test passes