fix: major bug in recording of chunks, files, packs
There were two related bugs when creating pack files and recording the
related metadata. If a single-chunk file happened to fill a pack file, it
would be recorded as part of the next pack record. This resulting in files
that were missing from the recorded pack.

A similar issue would result in a file record never being created, although
this happened very rarely.

cargo test passes