refactor: switch to EXAF for pack file format
The sodiumoxide library is deprecated and that raised a host of problems
with the current pack file format. After a month of development, the exaf-rs
crate will replace flate2, tar, and sodiumoxide crates. The pack file will
now have the random salt in its header, rather than in the pack record in
the database.

cargo test passes