Replace serde_cbor with ciborium
Replace the generated serde code that was used to serialize the entities to
the data store with hand-written code based on ciborium. This gives perfect
control over the process and allows for serde to be used on the entities
for other purposes, such as Leptos, which has different requirements than a
key/value data store.

cargo test passes